♻️ refactor server code
