+++ apps/server/src/common/constants/index.ts
@@ -1,0 +1,2 @@
+export const RESPONSE_SUCCESS = '200';
+export const RESPONSE_FAIL = '-1';
