+++ apps/server/src/common/enums/role.enum.ts
@@ -1,0 +1,13 @@
+/**
+ * 角色枚举
+ */
+export enum Role {
+  /**
+   * 管理员角色
+   */
+  ADMIN = 'admin',
+  /**
+   * 普通角色
+   */
+  NORMAL = 'normal',
+}
