+++ apps/server/src/guards/jwt.strategy.ts
