+++ apps/server/src/main.ts
@@ -2,7 +2,7 @@ import { NestFactory } from '@nestjs/core';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 
 import { AppModule } from './modules/app/app.module';
-import { appGlobalMiddleware } from './modules/app/useGlobal';
+import { appGlobalMiddleware } from './middlewares/global.middleware.ts';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
