+++ apps/server/src/modules/sms/sms.controller.ts
@@ -12,7 +12,7 @@ import {
   UpdatePhoneDto,
   VerifyPhoneDto,
 } from './dto/sms.dto';
-import { JwtAuthGuard } from '../auth/jwt/jwt.guard';
+import { JwtAuthGuard } from '../../guards/jwtAuth.guard';
 import { JwtPayloadDto } from '../auth/dto/jwt.dto';
 import { CallbackUserData } from '../auth/decorator/callback.decorator';
 import { string } from 'zod';
